lv_font.c: add draw_char()
  * compatible with picorv print_*() functions
  * rewrote UTF8 decoder to operate on 1 character at a time and keep internal state
  * add fast rectangles
  * add 2 font sizes to demo

+++ firmware/lib/frame_buffer.h
@@ -16,6 +16,6 @@ uint8_t getPixel(unsigned x, unsigned y);
 void fill(uint8_t shade);
 
 // Draw a rectangle from (x1, y1) to (x2, y2)
-void rect(uint8_t x1, uint8_t y1, uint8_t x2, uint8_t y2, uint8_t shade);
+void rect(int x1, int y1, int x2, int y2, uint8_t shade);
 
 #endif
